R4R: convert uint64 to int64 for BSON no support uint64

+++ modules/ibc/connect_open_init.go
@@ -9,7 +9,7 @@ type DocMsgConnectionOpenInit struct {
 	ClientId     string       `bson:"client_id"`
 	Counterparty Counterparty `bson:"counterparty"`
 	Version      Version      `bson:"version"`
-	DelayPeriod  uint64       `bson:"delay_period"`
+	DelayPeriod  int64        `bson:"delay_period"`
 	Signer       string       `bson:"signer"`
 }
 
@@ -20,7 +20,7 @@ func (m *DocMsgConnectionOpenInit) GetType() string {
 func (m *DocMsgConnectionOpenInit) BuildMsg(v interface{}) {
 	msg := v.(*MsgConnectionOpenInit)
 	m.ClientId = msg.ClientId
-	m.DelayPeriod = msg.DelayPeriod
+	m.DelayPeriod = int64(msg.DelayPeriod)
 	m.Signer = msg.Signer
 	if msg.Version != nil {
 		m.Version = Version{
